read 4 characteristics on mobile app

+++ Embedded Systems and Firmware/Self Monitor Stress System/GSR.c
@@ -13,8 +13,8 @@ void GSR_Init(){
     ADC0_Ch6_Init();
 }
 
-int8_t GSR_Read_Data(){
-    int gsr_average;
+int16_t GSR_Read_Data(){
+    int16_t gsr_average;
     gsr_average = ADC_Ch6_Samples(10);
     return gsr_average;
 }
